backend/requirements.txt
fastapi
uvicorn
python-dotenv
slowapi
google-generativeai
pydantic[email]
